capstone_movie_reco/requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
mlxtend
matplotlib
seaborn
plotly
networkx
sqlalchemy
pyyaml
scipy
statsmodels
tqdm
requests
jupyter
